refactor(app): extract video overlay derived helpers
