+++ packages/app/components/VideoPlayerOverlayImpl.tsx
@@ -29,28 +29,27 @@ import { getPlayerPageVideoHeight } from '@/lib/video-layout'
 import { useTabBarMetrics } from '@/lib/tabBarHeight'
 import { useCast } from '@/lib/cast'
 import { DevicePickerModal } from '@/components/cast'
+import {
+  computeMiniSize,
+  computeMiniBounds,
+  getCornerAnchors,
+  getMobileMiniPlayerSnapPosition,
+  resolveSnapTarget,
+  type MiniBounds,
+  type MiniPlayerCorner,
+  type Anchor,
+} from './video-player/overlayDerivedState'
 
 // Import modular video-player components
 import {
   // Constants
   MINI_PIP_MARGIN,
   MINI_PIP_CORNER_RADIUS,
-  MINI_PIP_COMPACT_WIDTH_FRACTION,
-  MINI_PIP_COMPACT_WIDTH_MIN,
-  MINI_PIP_COMPACT_WIDTH_MAX,
-  MINI_PIP_EXPANDED_WIDTH_FRACTION,
-  MINI_PIP_EXPANDED_WIDTH_MIN,
-  MINI_PIP_EXPANDED_WIDTH_MAX,
   TAB_BAR_HEIGHT,
 
   SPRING_CONFIG_BOUNCY,
   SPRING_CONFIG_TIGHT,
   SPRING_CONFIG_MINI_SNAP,
-  SNAP_LOW_SPEED,
-  SNAP_FLING_SPEED,
-  SNAP_TOSS_HORIZON,
-  SNAP_FLING_HORIZON,
-  SNAP_HYSTERESIS_PX,
   MINI_DRAG_SCALE,
   MINI_SHADOW_DOCKED,
   MINI_SHADOW_DRAGGING,
@@ -94,167 +93,7 @@ const Ionicons = ExpoIonicons
 const CHANNEL_META_CACHE_TTL_MS = 5 * 60 * 1000
 const channelMetaNameCache = new Map<string, { name: string | null; expiresAt: number }>()
 const ZERO_EDGE_INSETS = Object.freeze({ top: 0, right: 0, bottom: 0, left: 0 })
-type MiniPlayerCorner = 'bottom-right' | 'bottom-left' | 'top-right' | 'top-left'
-
-// ── Responsive mini-player geometry worklets ──────────────────────────
-
-interface MiniBounds {
-  leftBound: number
-  rightBound: number
-  topBound: number
-  bottomBound: number
-}
-
-interface Anchor { x: number; y: number; corner: MiniPlayerCorner }
-
-function computeMiniSize(screenWidth: number, aspectRatio: number, sizeMode: 'compact' | 'expanded' = 'compact') {
-  'worklet'
-  const ar = aspectRatio > 0 ? aspectRatio : 16 / 9
-  const fraction = sizeMode === 'expanded' ? MINI_PIP_EXPANDED_WIDTH_FRACTION : MINI_PIP_COMPACT_WIDTH_FRACTION
-  const minW = sizeMode === 'expanded' ? MINI_PIP_EXPANDED_WIDTH_MIN : MINI_PIP_COMPACT_WIDTH_MIN
-  const maxW = sizeMode === 'expanded' ? MINI_PIP_EXPANDED_WIDTH_MAX : MINI_PIP_COMPACT_WIDTH_MAX
-  const w = Math.max(minW, Math.min(maxW, Math.round(screenWidth * fraction)))
-  const h = Math.round(w / ar)
-  return { width: w, height: h }
-}
-
-function computeMiniBounds(
-  screenWidth: number,
-  screenHeight: number,
-  insetTop: number,
-  insetRight: number,
-  insetBottom: number,
-  insetLeft: number,
-  bottomChrome: number,
-  miniWidth: number,
-  miniHeight: number,
-): MiniBounds {
-  'worklet'
-  const margin = MINI_PIP_MARGIN
-  const bottomMargin = Math.max(insetBottom + margin, bottomChrome + margin)
-  return {
-    leftBound: insetLeft + margin,
-    rightBound: screenWidth - insetRight - margin - miniWidth,
-    topBound: insetTop + margin,
-    bottomBound: screenHeight - bottomMargin - miniHeight,
-  }
-}
-
-function getCornerAnchors(bounds: MiniBounds): [Anchor, Anchor, Anchor, Anchor] {
-  'worklet'
-  return [
-    { x: bounds.leftBound,  y: bounds.topBound,    corner: 'top-left' },
-    { x: bounds.rightBound, y: bounds.topBound,    corner: 'top-right' },
-    { x: bounds.leftBound,  y: bounds.bottomBound, corner: 'bottom-left' },
-    { x: bounds.rightBound, y: bounds.bottomBound, corner: 'bottom-right' },
-  ]
-}
-
-function resolveSnapTarget(
-  releaseX: number,
-  releaseY: number,
-  vx: number,
-  vy: number,
-  anchors: readonly Anchor[],
-  miniWidth: number,
-  miniHeight: number,
-  currentCorner: MiniPlayerCorner,
-  bounds: MiniBounds,
-): Anchor {
-  'worklet'
-  const speed = Math.sqrt(vx * vx + vy * vy)
-
-  // Projection horizon by velocity regime
-  let horizon = 0
-  if (speed >= SNAP_FLING_SPEED) {
-    horizon = SNAP_FLING_HORIZON
-  } else if (speed >= SNAP_LOW_SPEED) {
-    horizon = SNAP_TOSS_HORIZON
-  }
-
-  // Project release point, clamp to legal bounds
-  let targetX = releaseX + vx * horizon
-  let targetY = releaseY + vy * horizon
-  targetX = Math.max(bounds.leftBound, Math.min(bounds.rightBound, targetX))
-  targetY = Math.max(bounds.topBound, Math.min(bounds.bottomBound, targetY))
-
-  // Target card center
-  const tcx = targetX + miniWidth / 2
-  const tcy = targetY + miniHeight / 2
-
-  // Score each anchor by squared Euclidean center distance
-  let bestAnchor = anchors[0]
-  let bestScore = Infinity
-  let currentAnchorScore = Infinity
-
-  for (let i = 0; i < anchors.length; i++) {
-    const acx = anchors[i].x + miniWidth / 2
-    const acy = anchors[i].y + miniHeight / 2
-    const score = (acx - tcx) * (acx - tcx) + (acy - tcy) * (acy - tcy)
-    if (score < bestScore) {
-      bestScore = score
-      bestAnchor = anchors[i]
-    }
-    if (anchors[i].corner === currentCorner) {
-      currentAnchorScore = score
-    }
-  }
-
-  // Hysteresis: on slow release, prefer current corner if it's close enough
-  if (speed < SNAP_LOW_SPEED && currentAnchorScore < Infinity) {
-    const bestDist = Math.sqrt(bestScore)
-    const currentDist = Math.sqrt(currentAnchorScore)
-    if (currentDist - bestDist < SNAP_HYSTERESIS_PX) {
-      for (let i = 0; i < anchors.length; i++) {
-        if (anchors[i].corner === currentCorner) return anchors[i]
-      }
-    }
-  }
-
-  return bestAnchor
-}
-
-// ── JS-side snap position helper ──────────────────────────────────────
-
-function getMobileMiniPlayerSnapPosition({
-  corner,
-  screenWidth,
-  screenHeight,
-  topInset,
-  rightInset,
-  bottomInset,
-  leftInset,
-  bottomOffset,
-  aspectRatio,
-  sizeMode,
-}: {
-  corner: MiniPlayerCorner
-  screenWidth: number
-  screenHeight: number
-  topInset: number
-  rightInset: number
-  bottomInset: number
-  leftInset: number
-  bottomOffset: number
-  aspectRatio: number
-  sizeMode: 'compact' | 'expanded'
-}) {
-  const { width: miniWidth, height: miniHeight } = computeMiniSize(screenWidth, aspectRatio, sizeMode)
-  const bounds = computeMiniBounds(
-    screenWidth,
-    screenHeight,
-    topInset,
-    rightInset,
-    bottomInset,
-    leftInset,
-    bottomOffset,
-    miniWidth,
-    miniHeight,
-  )
-  const anchors = getCornerAnchors(bounds)
-  const anchor = anchors.find(a => a.corner === corner) ?? anchors[3] // default BR
-  return { x: anchor.x, y: anchor.y, width: miniWidth, height: miniHeight }
-}
+// ── Responsive mini-player geometry helpers live in ./video-player/overlayDerivedState ──
 
 export function VideoPlayerOverlay() {
   const insets = useContext(SafeAreaInsetsContext) ?? ZERO_EDGE_INSETS
